Some class were renamed
