+++ software/base/src/main/java/org/apache/brooklyn/entity/software/base/behavior/softwareprocess/SoftwareProcessImplBehaviorFactory.java
@@ -20,36 +20,13 @@ package org.apache.brooklyn.entity.software.base.behavior.softwareprocess;
 
 
 import org.apache.brooklyn.api.location.Location;
-import org.apache.brooklyn.entity.software.base.SoftwareProcess;
-import org.apache.brooklyn.entity.software.base.SoftwareProcessImpl;
 import org.apache.brooklyn.entity.software.base.behavior.softwareprocess.supplier.LocationFlagSupplier;
-import org.apache.brooklyn.entity.software.base.behavior.softwareprocess.supplier.PaasLocationFlagsSupplier;
 import org.apache.brooklyn.entity.software.base.lifecycle.LifecycleEffectorTasks;
-import org.apache.brooklyn.location.paas.PaasLocation;
 
-public class SoftwareProcessImplPaasBehaviourFactory
-        implements SoftwareProcessImplBehaviourFactory {
+public interface SoftwareProcessImplBehaviorFactory {
 
-    SoftwareProcessImpl entity;
-
-    public SoftwareProcessImplPaasBehaviourFactory(SoftwareProcessImpl entity){
-        this.entity = entity;
-    }
-
-    @Override
-    public LocationFlagSupplier getLocationFlagSupplier() {
-        return new PaasLocationFlagsSupplier(entity);
-    }
-
-    @Override
-    public LifecycleEffectorTasks getLifecycleEffectorTasks() {
-        return entity.getConfig(SoftwareProcess.PAAS_LIFECYCLE_EFFECTOR_TASKS);
-    }
-
-    @Override
-    public boolean isASupportedLocation(Location location) {
-        return location instanceof PaasLocation;
-    }
-
+    public LocationFlagSupplier getLocationFlagSupplier();
+    public LifecycleEffectorTasks getLifecycleEffectorTasks();
+    public boolean isASupportedLocation(Location location);
 
 }
